upgrade Vagrant to Ubuntu Focal and use bundler
Since the new GitHub Actions publish pipeline was added to
sync the website, it required adding Gemfile.lock in order
to cache gems between frequent runs.

The presence of the lock file changes the load behavior of
Jekyll, which requires use of Bundler (which historically
hasn't been needed in Vagrant environments).

So this changes the Vagrant environment to work more like
other environments, while also being upgraded to Jekyll 3.9
(latest github-pages version) for better compatibility with
Ruby 2.7 (in Focal).