Another attempt to fix node-sass with Node 16 build problem
Revert the previous commit disallowing the use of Node 16 as this is the
version used by yarn by default and so disabling its use just result in

error wxwidgets-website@1.2.0: The engine "node" is incompatible with
this module. Expected version ">= 10.13.0 <16". Got "16.13.0"

Try using node-sass >= 6.0.1, which is required for Node 16 support,
instead, and regenerate yarn.lock by simply removing the old file and
rerunning yarn.