Use Git for checking out wxWidgets sources instead of svn
Also get all the sources and not just the message catalogs and put them
in wxWidgets subdirectory.

Update the various scripts to use the files from there.

Also use v3 of the checkout action for checking out the main repository
itself too as a side effect (to avoid using inconsistent versions and
because there doesn't seem to be any reason to use the old v2 for the
newly added checkout step).